+++ src/integration/pages/idxPage/canonicalTests.js
@@ -1,9 +1,11 @@
+import runCrossPlatformTests from './crossPlatformTests';
 import {
   runCoreCanonicalTests,
   runCanonicalAnalyticsTests,
 } from '../../common';
 
 export default () => {
+  runCrossPlatformTests();
   runCoreCanonicalTests();
   runCanonicalAnalyticsTests();
 
@@ -16,32 +18,5 @@ export default () => {
     } else {
       expect(id).not.toBeInTheDocument();
     }
-  });
-
-  describe('Story Promo', () => {
-    it('I can see an image', () => {
-      const section = document.querySelector('section');
-
-      if (section) {
-        expect(section).toBeInTheDocument();
-
-        const imageEl = section.getElementsByTagName('img')[0];
-        expect(imageEl).toBeInTheDocument();
-        expect(imageEl).toMatchSnapshot();
-      }
-    });
-
-    it('I can see a headline', () => {
-      const section = document.querySelector('section');
-
-      if (section) {
-        expect(section).toBeInTheDocument();
-
-        const h3El = section.querySelector('h3');
-        expect(h3El).toBeInTheDocument();
-        expect(h3El.textContent).toBeTruthy();
-        expect(h3El.textContent).toMatchSnapshot();
-      }
-    });
   });
 };
