+++ src/integration/pages/idxPage/persian_afghanistan/amp.test.js
@@ -3,12 +3,6 @@
  * @pathname /persian/afghanistan
  */
 
-import runCrossPlatformTests from '../crossPlatformTests';
 import runAmpTests from '../ampTests';
 
-describe('AMP', () => {
-  describe(pageType, () => {
-    runCrossPlatformTests();
-    runAmpTests();
-  });
-});
+describe('AMP persian_afghanistan IDX page', runAmpTests);
