+++ src/integration/pages/idxPage/persian_afghanistan/canonical.test.js
@@ -3,12 +3,6 @@
  * @pathname /persian/afghanistan
  */
 
-import runCrossPlatformTests from '../crossPlatformTests';
 import runCanonicalTests from '../canonicalTests';
 
-describe('Canonical', () => {
-  describe(pageType, () => {
-    runCrossPlatformTests();
-    runCanonicalTests();
-  });
-});
+describe('Canonical persian_afghanistan IDX page', runCanonicalTests);
