+++ src/integration/pages/idxPage/ukraine_in_russian/amp.test.js
@@ -3,12 +3,6 @@
  * @pathname /ukrainian/ukraine_in_russian
  */
 
-import runCrossPlatformTests from '../crossPlatformTests';
 import runAmpTests from '../ampTests';
 
-describe('AMP', () => {
-  describe(pageType, () => {
-    runCrossPlatformTests();
-    runAmpTests();
-  });
-});
+describe('AMP ukraine_in_russian IDX page', runAmpTests);
