+++ src/integration/pages/idxPage/ukraine_in_russian/canonical.test.js
@@ -3,12 +3,6 @@
  * @pathname /ukrainian/ukraine_in_russian
  */
 
-import runCrossPlatformTests from '../crossPlatformTests';
 import runCanonicalTests from '../canonicalTests';
 
-describe('Canonical', () => {
-  describe(pageType, () => {
-    runCrossPlatformTests();
-    runCanonicalTests();
-  });
-});
+describe('Canonical ukraine_in_russian IDX page', runCanonicalTests);
